Refactor tokenizer a bit, implement whitespace
The tokenizer code is lazy, and it is written in terms of the input type
`Position` and the output type `Token`.

There are a few more token types to implement (literals, identifiers).

The only other thing that is seriously missing right now is graceful
error handling. This might take a bit of thinking, since it's not
obvious how it'll interact with the iterator-based lazy code.